% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMatrix.R
\name{PubMatrix}
\alias{PubMatrix}
\title{PubMatrix}
\usage{
PubMatrix(file, A = NULL, B = NULL, API.key = NULL,
  Database = "pubmed", daterange = NULL, outfile = NULL)
}
\arguments{
\item{file}{Path to the input file where two search lists are separated by comment character}

\item{A}{A vector of terms to search against B assuming file is not used}

\item{B}{A vector of terms to search against A assuming file is not used}

\item{API.key}{An API key obtained from Entrez; not necessary}

\item{Database}{Either pubmed or pmc}

\item{daterange}{a range of dates to search if desired}

\item{outfile}{A file path to export the search matrix as an .xlsx file.}
}
\value{
A search matrix
}
\description{
This function talleys boolean search results on PubMed using the Entrez search parameters
}
