% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMatrixR-package.R
\docType{package}
\name{PubMatrixR-package}
\alias{PubMatrixR}
\alias{PubMatrixR-package}
\title{PubMatrixR: Pairwise Literature Co-occurrence Matrices from NCBI Entrez}
\description{
`PubMatrixR` provides functions to query NCBI Entrez E-utilities for pairwise
co-occurrence counts across two term sets and to visualize the resulting
matrix-like data frame with heatmap helpers.
}
\section{Main functions}{

\itemize{
  \item [\code{\link{PubMatrix}}] Query NCBI and build a pairwise count matrix.
  \item [\code{\link{plot_pubmatrix_heatmap}}] Plot overlap percentages with clustering.
  \item [\code{\link{pubmatrix_heatmap}}] Convenience wrapper for heatmap plotting.
}
}

\section{Citation}{

See \code{citation("PubMatrixR")} for citation information.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ToledoEM/PubMatrixR-v2}
  \item Report bugs at \url{https://github.com/ToledoEM/PubMatrixR-v2/issues}
}

}
\author{
\strong{Maintainer}: Enrique Toledo \email{enriquetoledo@gmail.com} (\href{https://orcid.org/0000-0002-1460-4708}{ORCID})

Authors:
\itemize{
  \item Tyler Laird \email{tylerscottlaird@gmail.com}
}

}
\keyword{internal}
